requirements.txt
line-bot-sdk == 2.4.1
fastapi == 0.89.1
python-decouple == 3.6
uvicorn